backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
websockets==12.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
SQLAlchemy==2.0.25
aiosqlite==0.19.0
pydantic==2.5.3
python-dotenv==1.0.0
